Move graphql queries and mutations to separate file
