+++ src/screens/VisualTests/VisualTests.tsx
@@ -13,7 +13,7 @@ import { ProgressIcon } from "../../components/icons/ProgressIcon";
 import { Bar, Col, Row, Section, Sections, Text } from "../../components/layout";
 import { Text as CenterText } from "../../components/Text";
 import { GitInfoPayload, RunningBuildPayload } from "../../constants";
-import { getFragment, graphql } from "../../gql";
+import { getFragment } from "../../gql";
 import {
   AddonVisualTestsBuildQuery,
   AddonVisualTestsBuildQueryVariables,
@@ -26,205 +26,18 @@ import {
 import { UpdateStatusFunction } from "../../types";
 import { statusMap, testsToStatusUpdate } from "../../utils/testsToStatusUpdate";
 import { BuildProgress } from "./BuildProgress";
+import {
+  FragmentNextBuildFields,
+  FragmentStatusTestFields,
+  FragmentStoryBuildFields,
+  FragmentStoryTestFields,
+  MutationReviewTest,
+  QueryBuild,
+} from "./graphql";
 import { RenderSettings } from "./RenderSettings";
 import { SnapshotComparison } from "./SnapshotComparison";
 import { StoryInfo } from "./StoryInfo";
 import { Warnings } from "./Warnings";
-
-const QueryBuild = graphql(/* GraphQL */ `
-  query AddonVisualTestsBuild(
-    $projectId: ID!
-    $branch: String!
-    $gitUserEmailHash: String!
-    $slug: String
-    $storyId: String!
-    $testStatuses: [TestStatus!]!
-    $storyBuildId: ID!
-    $hasStoryBuildId: Boolean!
-  ) {
-    project(id: $projectId) {
-      name
-      lastBuild(
-        branches: [$branch]
-        slug: $slug
-        localBuilds: { localBuildEmailHash: $gitUserEmailHash }
-      ) {
-        ...NextBuildFields
-        ...StoryBuildFields @skip(if: $hasStoryBuildId)
-      }
-    }
-    storyBuild: build(id: $storyBuildId) @include(if: $hasStoryBuildId) {
-      ...StoryBuildFields
-    }
-  }
-`);
-
-const FragmentNextBuildFields = graphql(/* GraphQL */ `
-  fragment NextBuildFields on Build {
-    __typename
-    id
-    commit
-    committedAt
-    browsers {
-      id
-      key
-      name
-    }
-    ... on StartedBuild {
-      changeCount: testCount(results: [ADDED, CHANGED, FIXED])
-      brokenCount: testCount(results: [CAPTURE_ERROR])
-      testsForStatus: tests(first: 1000, statuses: $testStatuses) {
-        nodes {
-          ...StatusTestFields
-        }
-      }
-    }
-    ... on CompletedBuild {
-      result
-      changeCount: testCount(results: [ADDED, CHANGED, FIXED])
-      brokenCount: testCount(results: [CAPTURE_ERROR])
-      testsForStatus: tests(statuses: $testStatuses) {
-        nodes {
-          ...StatusTestFields
-        }
-      }
-    }
-  }
-`);
-
-const FragmentStoryBuildFields = graphql(/* GraphQL */ `
-  fragment StoryBuildFields on Build {
-    __typename
-    id
-    number
-    branch
-    uncommittedHash
-    status
-    ... on StartedBuild {
-      startedAt
-      testsForStory: tests(storyId: $storyId) {
-        nodes {
-          ...StoryTestFields
-        }
-      }
-    }
-    ... on CompletedBuild {
-      startedAt
-      testsForStory: tests(storyId: $storyId) {
-        nodes {
-          ...StoryTestFields
-        }
-      }
-    }
-  }
-`);
-
-const FragmentStatusTestFields = graphql(/* GraphQL */ `
-  fragment StatusTestFields on Test {
-    id
-    status
-    story {
-      storyId
-    }
-  }
-`);
-
-const FragmentStoryTestFields = graphql(/* GraphQL */ `
-  fragment StoryTestFields on Test {
-    id
-    status
-    result
-    webUrl
-    comparisons {
-      id
-      result
-      browser {
-        id
-        key
-        name
-        version
-      }
-      captureDiff {
-        diffImage {
-          imageUrl
-          imageWidth
-        }
-      }
-      headCapture {
-        captureImage {
-          imageUrl
-          imageWidth
-        }
-        captureError {
-          kind
-          ... on CaptureErrorInteractionFailure {
-            error
-          }
-          ... on CaptureErrorJSError {
-            error
-          }
-          ... on CaptureErrorFailedJS {
-            error
-          }
-        }
-      }
-      baseCapture {
-        captureImage {
-          imageUrl
-          imageWidth
-        }
-      }
-      viewport {
-        id
-        name
-        width
-        isDefault
-      }
-    }
-    parameters {
-      viewport {
-        id
-        name
-        width
-        isDefault
-      }
-    }
-    story {
-      storyId
-      name
-      component {
-        name
-      }
-    }
-  }
-`);
-
-const MutationReviewTest = graphql(/* GraphQL */ `
-  mutation ReviewTest($input: ReviewTestInput!) {
-    reviewTest(input: $input) {
-      updatedTests {
-        id
-        status
-      }
-      userErrors {
-        ... on UserError {
-          __typename
-          message
-        }
-        ... on BuildSupersededError {
-          build {
-            id
-          }
-        }
-        ... on TestUnreviewableError {
-          test {
-            id
-          }
-        }
-      }
-    }
-  }
-`);
 
 const createEmptyStoryStatusUpdate = (state: API_StatusState) => {
   return Object.fromEntries(Object.entries(state).map(([id, update]) => [id, null]));
